requirements.txt
Django==2.2
djangorestframework==3.12.2
Pillow==8.0.1